refactor: move tools to their own folder
