+++ src/tools/get-book-by-title/index.ts
@@ -6,10 +6,7 @@ import {
 import axios from "axios";
 import { z } from "zod";
 
-import {
-  BookInfo,
-  OpenLibrarySearchResponse,
-} from "./get-book-by-title.types.js";
+import { BookInfo, OpenLibrarySearchResponse } from "./types.js";
 
 // Schema for the get_book_by_title tool arguments
 export const GetBookByTitleArgsSchema = z.object({
