+++ src/tools/get-book-cover/index.test.ts
@@ -1,7 +1,7 @@
 import { McpError, ErrorCode } from "@modelcontextprotocol/sdk/types.js";
 import { describe, expect, it } from "vitest";
 
-import { handleGetBookCover } from "./get-book-cover.js";
+import { handleGetBookCover } from "./index.js";
 
 describe("handleGetBookCover", () => {
   it("should generate correct URL with ISBN", async () => {
