+++ src/tools/get-book-cover/index.ts
@@ -1,12 +1,8 @@
-import {
-  CallToolResult,
-  McpError,
-  ErrorCode,
-} from "@modelcontextprotocol/sdk/types.js";
+import { McpError, ErrorCode } from "@modelcontextprotocol/sdk/types.js";
 import { z } from "zod";
 
 // Schema for the get_book_cover tool arguments
-const GetBookCoverArgsSchema = z.object({
+export const GetBookCoverArgsSchema = z.object({
   key: z.enum(["ISBN", "OCLC", "LCCN", "OLID", "ID"], {
     errorMap: () => ({
       message: "Key must be one of ISBN, OCLC, LCCN, OLID, ID",
@@ -19,7 +15,7 @@ const GetBookCoverArgsSchema = z.object({
     .transform((val) => val || "L"),
 });
 
-const handleGetBookCover = async (args: unknown): Promise<CallToolResult> => {
+const handleGetBookCover = async (args: unknown) => {
   const parseResult = GetBookCoverArgsSchema.safeParse(args);
 
   if (!parseResult.success) {
@@ -33,6 +29,7 @@ const handleGetBookCover = async (args: unknown): Promise<CallToolResult> => {
   }
 
   const { key, value, size } = parseResult.data;
+
   // Construct the URL according to the Open Library Covers API format
   const coverUrl = `https://covers.openlibrary.org/b/${key.toLowerCase()}/${value}-${size}.jpg`;
