--- src/types.ts
@@ -1,1 +1,0 @@
-// Removed the DetailedAuthorInfo interface as per the code change suggestion. The file is now empty.
